GITHUB-16: fix GCP build script
There was an invalid model variable left in a build script that caused failures.
Apart from that, there was removed a startup script which clones the repo to /tmp folder, as there are lots of permission issues occur after ssh'ing with a specific user. It'd make more sense to let a user clone the repo into their home folder after deployment.